make ipython executor dependency optional
